User and Tag policies

+++ app/Http/Controllers/TagController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\tag;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class TagController extends Controller
 {
@@ -82,6 +83,7 @@ class TagController extends Controller
      */
     public function edit(Tag $tag)
     {
+        abort_unless(Gate::allows('update', $tag), 403);
         return view('tag.edit')->with([
             'tag' => $tag
         ]);
@@ -96,6 +98,8 @@ class TagController extends Controller
      */
     public function update(Request $request, Tag $tag)
     {
+
+        abort_unless(Gate::allows('update', $tag), 403);
         $request->validate([
             'name' => 'required|min:3',
             'style' => 'required|min:3',
@@ -121,6 +125,9 @@ class TagController extends Controller
      */
     public function destroy(Tag $tag)
     {
+
+        abort_unless(Gate::allows('delete', $tag), 403);
+
         $OldTag = $tag->name;
         $tag->delete();
